adjusted resources and create the initial module
