+++ az/concrete-cms/inputs.yaml
@@ -1,0 +1,4 @@
+---
+resource_group_name: example
+resource_group_name_override: ""
+tags: {}
