+++ az/env.yaml
@@ -1,11 +1,10 @@
 ---
-environment: example
+environment: dev
 subscription: 00000000-0000-0000-0000-000000000000
 labels:
-  deployment: kped
+  deployment: con
   environment: example
   owner: kpeder
   team: devops
-locations:
-  primary: us-east1
-  secondary: us-central1
+location: us-east
+location_short: eus
