--- az/resource-group/inputs.yaml
@@ -1,46 +1,0 @@
----
-audit_log_config:
-  - service: "allServices"
-    log_type: "DATA_READ"
-    exempted_members: []
-  - service: "allServices"
-    log_type: "DATA_WRITE"
-    exempted_members: []
-  - service: "allServices"
-    log_type: "ADMIN_READ"
-    exempted_members: []
-
-folder:
-  names:
-    - "folder"
-  parent:
-    id: ""
-    type: "" # can be one of 'organizations' or 'folders', depending on where the deployment should be rooted
-
-project:
-  activate_apis:
-    - "cloudbuild.googleapis.com"
-    - "cloudfunctions.googleapis.com"
-    - "cloudresourcemanager.googleapis.com"
-    - "compute.googleapis.com"
-    - "container.googleapis.com"
-    - "iam.googleapis.com"
-    - "iap.googleapis.com"
-    - "logging.googleapis.com"
-    - "monitoring.googleapis.com"
-    - "pubsub.googleapis.com"
-    - "run.googleapis.com"
-    - "servicenetworking.googleapis.com"
-    - "serviceusage.googleapis.com"
-    - "storage-api.googleapis.com"
-    - "storage-component.googleapis.com"
-  auto_create_network: false
-  billing_account_override: ""
-  default_service_account: "deprivilege"
-  folder_id_override: ""
-  project_id_override: ""
-  random_project_id: true
-  state_bucket:
-    force_destroy: false
-    uniform_access: true
-    versioning: true
